requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
paramiko>=3.4.0
psutil>=5.9.0
